Add authorization to notes service
- Add User property to Note entity, update Create method, update NoteCreatedEvent
- Update ef core configuration, create migration
- Create IUsersService and implementation using gRPC
- Add Authentication options, register authentication in DI
- Add UserRequest and UserRequestBehavior
- Change AddNote, GetNote, GetNoteFromRoot, RemoveNote, UpdateNote to inherit its request from UserRequest
- Add UserNotFoundException and UnhandledErrorException
- Add Authorize filter on CreateNote, DeleteNote, UpdateNote, GetNote, GetNotesFromRoot in NotesController
- Update seed data and tests

+++ Np.NotesService/Np.NotesService.Application/Notes/RemoveNote/RemoveNoteCommandHandler.cs
@@ -22,12 +22,15 @@ namespace Np.NotesService.Application.Notes.RemoveNote
         public async Task<Result> Handle(RemoveNoteCommand request, CancellationToken cancellationToken)
         {
             var note = await _notesRepository.GetNoteById(request.NoteId);
-
-            if (note == null) 
+            if (note == null)
             {
                 return Result.Success();
             }
 
+            if (!note.User.Id.Equals(request.UserId!.Value))
+            {
+                return Result.Success();
+            }
             _notesRepository.Delete(note);
             note.AddDomainEvent(new NoteRemovedEvent(note.Id));
